Update  Issue 5177 Status: NeedsReview cc: adblockforchrome
This is the initial implementation and would like to receive feedbacks and comments on how to improve this. Currently, unscribeall is slow when there are a lot of filters subscribed. Also, confirmation for excessive subscriptions is not yet fixed. I commented out the original implementation as a basis and to serve as a guide.

To test, open options page for adblock and see the description if the expected result is met.